pyproject.toml
[project]
name = "demo-legal-rag"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "absl-py>=2.3.1",
    "chonkie[tokenizers]>=1.0.0",
    "colabtools>=0.0.1",
    "datasets>=4.4.1",
    "dspy>=3.0.4",
    "etils[colab,eapp,ecolab,edc,epath,epy]>=1.13.0",
    "google-colab",
    "grain>=0.2.14",
    "huggingface-hub>=0.36.0",
    "jinja2>=3.1.6",
    "pydrive2>=1.21.3",
    "pylate>=1.3.4",
    "rich[jupyter]>=14.2.0",
    "sentencepiece>=0.2.1",
    "tokenizers>=0.22.1",
    "toolz>=1.1.0",
    "transformers>=4.38,<5",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[[tool.uv.index]]
url = "https://flashinfer.ai/whl/cu128"

[tool.uv.sources]
google-colab = { git = "https://github.com/googlecolab/colabtools" }

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "jupytext>=1.18.1",
    "pdbp>=1.8.1",
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.ruff]
extend-exclude = ["demo.ipynb"]
